Add import tenant functionality (#54)
* chore: use proper name for var

* fix: make sure tenant name takes precedence

* feat: remove defaulter from account

* feat: make sure tenantclass is not reconciled on status changes

* feat: WIP tenant import

* feat: work out import feature

* fix: some import reconcilitation improvements

* feat: finalize import feature

* chore: make linter an testcases happy again

* feat: update docs and remove phaseRetaining as a whole

* fix: minor issues and prettyfying

+++ api/v1alpha1/s3bucket_types.go
@@ -100,7 +100,7 @@ type S3BucketStatus struct {
 	Phase BucketPhase `json:"phase,omitempty"`
 
 	// DrainStatus tracks active drain operation progress.
-	// Only present during active drain. Cleared when drain completes or is cancelled.
+	// Only present during active drain. Cleared when drain completes or is canceled.
 	// +optional
 	DrainStatus *BucketDrainStatus `json:"drainStatus,omitempty"`
 
@@ -149,7 +149,7 @@ type BucketDrainStatus struct {
 	IsDeletingObjects bool `json:"isDeletingObjects"`
 
 	// InitialObjectCount is the number of objects when drain started.
-	InitialObjectCount int64 `json:"initialObjectCount,omitempty"`
+	InitialObjectCount int32 `json:"initialObjectCount,omitempty"`
 
 	// InitialObjectBytes is the total bytes when drain started.
 	InitialObjectBytes int64 `json:"initialObjectBytes,omitempty"`
